requirements.txt
### Infrastructure
uvicorn==0.23.2
pydantic_settings==2.0.3
python-dotenv
pydantic
PyYAML==6.0.1
python-dotenv==1.0.0

### Functional
fastapi==0.104.0
grpcio
mock

### Operations
starlette-context==0.3.6
importlib-metadata==7.0.1